Split tests to modules keep 1:1 ratio
